Implement resolve also for repositories
