+++ src/main/java/io/vertx/json/schema/JsonSchema.java
@@ -112,10 +112,10 @@ public interface JsonSchema {
   /**
    * Tries to resolve all internal references. External references are not resolved.
    *
-   * The result is an object where all references have been resolved. Resolution of references is shallow This
-   * should normally not be a problem for this use case.
+   * The result is an object where all references have been resolved. Resolution of references is shallow.
+   * This should normally not be a problem for this use case.
    *
-   * @return a new {@link JsonObject} representing the schema with {@code $refs} replaced by their value.
+   * @return a new {@link JsonObject} representing the schema with {@code $ref}s replaced by their value.
    * @throws SchemaException when the resolution is impossible. One of such cases is circular referencing.
    */
   JsonObject resolve();
