+++ src/test/resources/resolve/pet.model.json
@@ -1,0 +1,99 @@
+{
+  "Pet": {
+    "required": [
+      "name",
+      "photoUrls"
+    ],
+    "type": "object",
+    "properties": {
+      "id": {
+        "type": "integer",
+        "format": "int64"
+      },
+      "category": {
+        "$ref": "#/Category"
+      },
+      "name": {
+        "type": "string",
+        "example": "doggie"
+      },
+      "photoUrls": {
+        "type": "array",
+        "xml": {
+          "name": "photoUrl",
+          "wrapped": true
+        },
+        "items": {
+          "type": "string"
+        }
+      },
+      "tags": {
+        "type": "array",
+        "xml": {
+          "name": "tag",
+          "wrapped": true
+        },
+        "items": {
+          "$ref": "#/Tag"
+        }
+      },
+      "status": {
+        "type": "string",
+        "description": "pet status in the store",
+        "enum": [
+          "available",
+          "pending",
+          "sold"
+        ]
+      }
+    },
+    "xml": {
+      "name": "Pet"
+    }
+  },
+  "Category": {
+    "type": "object",
+    "properties": {
+      "id": {
+        "type": "integer",
+        "format": "int64"
+      },
+      "name": {
+        "type": "string"
+      }
+    },
+    "xml": {
+      "name": "Category"
+    }
+  },
+  "Tag": {
+    "type": "object",
+    "properties": {
+      "id": {
+        "type": "integer",
+        "format": "int64"
+      },
+      "name": {
+        "type": "string"
+      }
+    },
+    "xml": {
+      "name": "Tag"
+    }
+  },
+  "ApiResponse": {
+    "type": "object",
+    "properties": {
+      "code": {
+        "type": "integer",
+        "format": "int32"
+      },
+      "type": {
+        "type": "string"
+      },
+      "message": {
+        "type": "string"
+      }
+    }
+  }
+}
