+++ src/test/resources/resolve/store.api.json
@@ -1,0 +1,155 @@
+{
+  "store-inventory": {
+    "get": {
+      "tags": [
+        "store"
+      ],
+      "summary": "Returns pet inventories by status",
+      "description": "Returns a map of status codes to quantities",
+      "operationId": "getInventory",
+      "responses": {
+        "200": {
+          "description": "successful operation",
+          "content": {
+            "application/json": {
+              "schema": {
+                "type": "object",
+                "additionalProperties": {
+                  "type": "integer",
+                  "format": "int32"
+                }
+              }
+            }
+          }
+        }
+      },
+      "security": [
+        {
+          "api_key": []
+        }
+      ]
+    }
+  },
+  "store-order": {
+    "post": {
+      "tags": [
+        "store"
+      ],
+      "summary": "Place an order for a pet",
+      "operationId": "placeOrder",
+      "requestBody": {
+        "description": "order placed for purchasing the pet",
+        "content": {
+          "*/*": {
+            "schema": {
+              "$ref": "store.model.json#/Order"
+            }
+          }
+        },
+        "required": true
+      },
+      "responses": {
+        "200": {
+          "description": "successful operation",
+          "content": {
+            "application/xml": {
+              "schema": {
+                "$ref": "store.model.json#/Order"
+              }
+            },
+            "application/json": {
+              "schema": {
+                "$ref": "store.model.json#/Order"
+              }
+            }
+          }
+        },
+        "400": {
+          "description": "Invalid Order",
+          "content": {}
+        }
+      },
+      "x-codegen-request-body-name": "body"
+    }
+  },
+  "store-order-orderId": {
+    "get": {
+      "tags": [
+        "store"
+      ],
+      "summary": "Find purchase order by ID",
+      "description": "For valid response try integer IDs with value >= 1 and <= 10.         Other\nvalues will generated exceptions",
+      "operationId": "getOrderById",
+      "parameters": [
+        {
+          "name": "orderId",
+          "in": "path",
+          "description": "ID of pet that needs to be fetched",
+          "required": true,
+          "schema": {
+            "maximum": 10,
+            "minimum": 1,
+            "type": "integer",
+            "format": "int64"
+          }
+        }
+      ],
+      "responses": {
+        "200": {
+          "description": "successful operation",
+          "content": {
+            "application/xml": {
+              "schema": {
+                "$ref": "store.model.json#/Order"
+              }
+            },
+            "application/json": {
+              "schema": {
+                "$ref": "store.model.json#/Order"
+              }
+            }
+          }
+        },
+        "400": {
+          "description": "Invalid ID supplied",
+          "content": {}
+        },
+        "404": {
+          "description": "Order not found",
+          "content": {}
+        }
+      }
+    },
+    "delete": {
+      "tags": [
+        "store"
+      ],
+      "summary": "Delete purchase order by ID",
+      "description": "For valid response try integer IDs with positive integer value.         Negative\nor non-integer values will generate API errors",
+      "operationId": "deleteOrder",
+      "parameters": [
+        {
+          "name": "orderId",
+          "in": "path",
+          "description": "ID of the order that needs to be deleted",
+          "required": true,
+          "schema": {
+            "minimum": 1,
+            "type": "integer",
+            "format": "int64"
+          }
+        }
+      ],
+      "responses": {
+        "400": {
+          "description": "Invalid ID supplied",
+          "content": {}
+        },
+        "404": {
+          "description": "Order not found",
+          "content": {}
+        }
+      }
+    }
+  }
+}
