+++ src/test/resources/resolve/store.model.json
@@ -1,0 +1,39 @@
+{
+  "Order": {
+    "type": "object",
+    "properties": {
+      "id": {
+        "type": "integer",
+        "format": "int64"
+      },
+      "petId": {
+        "type": "integer",
+        "format": "int64"
+      },
+      "quantity": {
+        "type": "integer",
+        "format": "int32"
+      },
+      "shipDate": {
+        "type": "string",
+        "format": "date-time"
+      },
+      "status": {
+        "type": "string",
+        "description": "Order Status",
+        "enum": [
+          "placed",
+          "approved",
+          "delivered"
+        ]
+      },
+      "complete": {
+        "type": "boolean",
+        "default": false
+      }
+    },
+    "xml": {
+      "name": "Order"
+    }
+  }
+}
