+++ src/test/resources/resolve/user.model.json
@@ -1,0 +1,37 @@
+{
+  "User": {
+    "type": "object",
+    "properties": {
+      "id": {
+        "type": "integer",
+        "format": "int64"
+      },
+      "username": {
+        "type": "string"
+      },
+      "firstName": {
+        "type": "string"
+      },
+      "lastName": {
+        "type": "string"
+      },
+      "email": {
+        "type": "string"
+      },
+      "password": {
+        "type": "string"
+      },
+      "phone": {
+        "type": "string"
+      },
+      "userStatus": {
+        "type": "integer",
+        "description": "User Status",
+        "format": "int32"
+      }
+    },
+    "xml": {
+      "name": "User"
+    }
+  }
+}
